requirements.txt
cryptography
PyYAML
requests